# [FLINK-32594][runtime]Support EndOfStreamTrigger and isOutputOnEOF operator attribute to optimize task deployment