--- a/flink-runtime/src/main/java/org/apache/flink/runtime/checkpoint/CheckpointCoordinatorDeActivator.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/checkpoint/CheckpointCoordinatorDeActivator.java
@@ -20,7 +20,11 @@ package org.apache.flink.runtime.checkpoint;
 
 import org.apache.flink.api.common.JobID;
 import org.apache.flink.api.common.JobStatus;
+import org.apache.flink.runtime.executiongraph.ExecutionJobVertex;
 import org.apache.flink.runtime.executiongraph.JobStatusListener;
+import org.apache.flink.runtime.jobgraph.JobVertexID;
+
+import java.util.Map;
 
 import static org.apache.flink.util.Preconditions.checkNotNull;
 
@@ -31,19 +35,31 @@ import static org.apache.flink.util.Preconditions.checkNotNull;
 public class CheckpointCoordinatorDeActivator implements JobStatusListener {
 
     private final CheckpointCoordinator coordinator;
+    private final Map<JobVertexID, ExecutionJobVertex> tasks;
 
-    public CheckpointCoordinatorDeActivator(CheckpointCoordinator coordinator) {
+    public CheckpointCoordinatorDeActivator(
+            CheckpointCoordinator coordinator, Map<JobVertexID, ExecutionJobVertex> tasks) {
         this.coordinator = checkNotNull(coordinator);
+        this.tasks = checkNotNull(tasks);
     }
 
     @Override
     public void jobStatusChanges(JobID jobId, JobStatus newJobStatus, long timestamp) {
-        if (newJobStatus == JobStatus.RUNNING) {
-            // start the checkpoint scheduler
+        if (newJobStatus == JobStatus.RUNNING && allTasksOutputPipelinedOrPipelinedBounded()) {
+            // start the checkpoint scheduler if there are all pipeline tasks
             coordinator.startCheckpointScheduler();
         } else {
             // anything else should stop the trigger for now
             coordinator.stopCheckpointScheduler();
         }
+    }
+
+    private boolean allTasksOutputPipelinedOrPipelinedBounded() {
+        for (Map.Entry<JobVertexID, ExecutionJobVertex> task : tasks.entrySet()) {
+            if (!task.getValue().getJobVertex().allOutputsPipelinedOrPipelinedBounded()) {
+                return false;
+            }
+        }
+        return true;
     }
 }