@RanJinh 2023-07-17
I added the `org.apache.flink.streaming.api.windowing.assigners.EndOfStreamWindows.assignWindows` in the `archunit-violations` file to avid `org.apache.flink.architecture.ArchitectureTest` detecting the annotation, which also be done for other window assigners.