@RanJinh 2023-07-24
@flinkbot run azure