@RanJinh 2023-07-25
@flinkbot run azure