@RanJinh 2023-07-26
@flinkbot run azure